refactor: use .story.tsx naming convention
