+++ src/components/Loading/Loading.story.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import centered from '@storybook/addon-centered';
 
-import { Loading } from './';
+import { Loading } from '.';
 
 storiesOf('Loading', module)
   .addDecorator(centered)
