+++ src/components/SearchBar/SearchBar.story.tsx
